Use standard header locations for poll.h and signal.h
sys/poll.h and sys/signal.h just redirect to the standardized
location for these headers, poll.h and signal.h.

On musl libc, including the incorrect header path results in a
warning, and since breakpad is built with -Werror, this is an error.

In exception_handler.cc, signal.h is already included earlier, so
we can drop the sys/signal.h include.

Bug: google-breakpad:631
Change-Id: If36d326453e3267d38a5b92ed1301f828e46befe