Update patch set 1
Patch Set 1:

> i haven't fully double checked the structs/offsets, but i don't think this works.  if you look at ucontext_constants.h, you can see that it's hardcoded to use offsets specific to Android.
> 
> the ucontext structures are pretty ugly.  each kernel abi has a struct/layout for its syscall interfaces which will be the same between C libraries (bionic/glibc/musl/etc...), but then each C library has its own struct/layout.  some might try to re-use the same layout to simplify, but there's no requirement.

The constants match between Android and non-Android Linux, except for the offset of __fpregs_mem on i386 and x86_64, which I made conditional. The offsets are made sure to match ucontext_t by breakpad_getcontext_unittest.cc, which I ran on x86_64, i686, and aarch64. I will double check the constants on 32-bit ARM and MIPS, since I don't have the hardware to run the tests.

I'm not sure there is a way to avoid hard-coding the constants, except maybe moving the breakpad_getcontext implementation to a C source with inline assembly.

> sys/ucontext.h provides ucontext_t which are the C library definitions/layouts, not the kernel ones.  the kernel definitions are just "struct ucontext".

Isn't it just the libc constants that matter, since getcontext is used with ucontext_t in exception_handler.cc?

Patch-set: 1