Update patch set 3
Patch Set 3:

(1 comment)

i remember now the other big problem i have with this.  POSIX does not specify the C library layout of the ucontext_t structure.  it only requires that a few members exist in ucontext_t, but that's it -- nothing about order, padding, or sizes of individual members.

> https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/signal.h.html
> The <signal.h> header shall define the ucontext_t type as a structure that shall include at least the following members:
> 
> * ucontext_t *uc_link     Pointer to the context that is resumed when this context returns.
> * sigset_t    uc_sigmask  The set of signals that are blocked when this context is active.
> * stack_t     uc_stack    The stack used by this context.
> * mcontext_t  uc_mcontext A machine-specific representation of the saved context.

further, mcontext_t isn't defined anywhere in POSIX (anymore).

so the current constants header is written specifically with bionic in mind.  if other C libs (like musl) happen to copy the same struct around, it'll work, but that's kind of a weak guarantee.  you can see that already with the additions of __ANDROID__ you have in there where you had to adjust it slightly.

if we wanted to truly make this portable, we could use a lot of AC_COMPUTE_INT calls with offsetof.  but i can see that being a bit painful.  how about we just add an explicit guard to the top of the file kind of like what we have with the arch #error ?

 #if !defined(__BIONIC__) && !defined(__musl__)
 #error "This header has not been ported for your C library"
 #endif

that would also avoid needing the HAVE_GETCONTEXT check in it.

Patch-set: 3
Reviewer: Gerrit User 1000768 <1000768@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>