Update patch set 4
Patch Set 4:

(1 comment)

> i remember now the other big problem i have with this.  POSIX does not specify the C library layout of the ucontext_t structure.  it only requires that a few members exist in ucontext_t, but that's it -- nothing about order, padding, or sizes of individual members.
> 
> further, mcontext_t isn't defined anywhere in POSIX (anymore).

Indeed, but this is a Linux-specific implementation, not something meant to be portable across all POSIX systems.

> so the current constants header is written specifically with bionic in mind.  if other C libs (like musl) happen to copy the same struct around, it'll work, but that's kind of a weak guarantee.  you can see that already with the additions of __ANDROID__ you have in there where you had to adjust it slightly.

I did some more investigation, and found that for signal handlers registered with SA_SIGINFO, the kernel sets up a ucontext_t and stack frame when it calls the signal handler. So it is not just coincidence that the layout of public members of ucontext_t and mcontext_t match across all Linux libcs (bionic, glibc, musl, uclibc, dietlibc) and the kernel itself. (Theoretically, I suppose a libc could wrap all signal handlers, to translate to its own struct before calling the real handler, but I'm not aware of any that does that.)

The only difference here is where the storage pointed to by uc_mcontext.fpregs resides. For ucontext_t obtained from a signal handler, I've found that fpregs doesn't even line up with __fpregs_mem on bionic, glibc, or musl; the kernel allocates the FP state on the stack separately from the struct ucontext. The ucontext_t __fpregs_mem storage must just be to support getcontext(), so it has a place to store the FP state. All that matters as far as breakpad is concerned is that it doesn't overlap with uc_mcontext, lies within the ucontext_t, and is aligned correctly.

In any case, I believe the current proposed code with the single conditional for __fpregs_mem works with all Linux libcs I'm aware of.

> if we wanted to truly make this portable, we could use a lot of AC_COMPUTE_INT calls with offsetof.  but i can see that being a bit painful.

I don't think this would be acceptable for most downstream projects using breakpad, since they would have to replicate the same offset calculation tests in their build systems.

> how about we just add an explicit guard to the top of the file kind of like what we have with the arch #error ?
> 
>  #if !defined(__BIONIC__) && !defined(__musl__)
>  #error "This header has not been ported for your C library"
>  #endif
> 
> that would also avoid needing the HAVE_GETCONTEXT check in it.

musl does not have an identifying __musl__ define, so this isn't an option.

Patch-set: 4