Update patch set 4
Patch Set 4: Code-Review+1

> I did some more investigation, and found that for signal handlers registered with SA_SIGINFO, the kernel sets up a ucontext_t and stack frame when it calls the signal handler.

thanks for catching that.  POSIX is pretty clear here:

> https://pubs.opengroup.org/onlinepubs/9699919799/functions/sigaction.html
> the third argument can be cast to a pointer to an object of type ucontext_t

so that provides the strong binding between the ucontext_t as defined in signal.h by the C library to the struct ucontext that the kernel is using when it calls the signal handler in userspace.

> In any case, I believe the current proposed code with the single conditional for __fpregs_mem works with all Linux libcs I'm aware of.

the fact we need a check at all means that it's not reliable.  however, the unittests already enforce the offsets are correct at compile time, and exist within the struct, so that should provide the cover we need.  if people are trying to port to a new C library but don't build+run the unittests, that's not our problem.

> I don't think this would be acceptable for most downstream projects using breakpad, since they would have to replicate the same offset calculation tests in their build systems.

hmm, along those lines, if your goal is to port Chromium to breakpad, we might have trouble.  Chromium has its own BUILD.gn rules for building breakpad and they'll be updated to always define HAVE_GETCONTEXT since we target glibc.

Patch-set: 4
Reviewer: Gerrit User 1000768 <1000768@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1