Update patch set 5
Patch Set 5:

(1 comment)

Thanks again for the review.

> hmm, along those lines, if your goal is to port Chromium to breakpad, we might have trouble.  Chromium has its own BUILD.gn rules for building breakpad and they'll be updated to always define HAVE_GETCONTEXT since we target glibc.

The motivation for this was for use in firefox, whose build system does support checking for functions during configure, so there will be no problem there.

I was hoping that chromium's GN scripts would have something equivalent to AC_CHECK_FUNC, but it looks like it doesn't. At the very least, it will be much nicer for musl-based distributions to be able to simply comment out the HAVE_GETCONTEXT define. An `is_musl` GN argument is another possibility.

Patch-set: 5