Convert a live coredump to a minidump.
While the kernel generates and streams the coredump, convert it to a
minidump, taking advantage of access to /proc to avoid buffering all of
the coredump.

This is to be used by a utility launched from
/proc/sys/kernel/core_pattern.  This is not intended for use from the
exception handler of the crashing program itself, and so it may
dynamically allocate memory on the heap.

Bug: internal b/153016480
Change-Id: Idf3cb3399edd76d92645812055ed719653762e82