Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Hi. This is all deprecated code and has been superseded by Crashpad.
> 
> I wish you would have engaged me sooner, I would have encouraged you to do this in Crashpad instead. Having a Crashpad handler, or auxiliary program that speaks to the Crashpad handler, launched out of the core_pattern pipe, is something that we’d be interested in there.
> 
> https://crashpad.chromium.org

I have a proof-of-concept implementation for crashpad too, and just posted about it to the crashpad dev mailing list.  It just isn't working for me.  Here is a link to that CL: https://chromium-review.googlesource.com/c/chromium/src/+/2161657

This Breakpad CL was useful to me to test the idea and prove that it works.

Patch-set: 4