Update patch set 4
Patch Set 4:

(1 comment)

This is great.

The change that broke this was made to fix a test. Do all of the tests still pass?

The fact that the change that broke this was able to be made must be that there was no test that covers this particular case. Can you add a test that makes sure that this behavior is working correctly?

Patch-set: 4
Reviewer: Gerrit User 1174099 <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>