Delete formerly leaked objects
Module documents that it takes ownership of the Function pointers that
are passed to AddFunction. However, when the Function's address is
outside the module's bounds, it doesn't save the pointer and it leaks
leaks. Similarly with Externs and StackFrames.

This patch fixes the leak by having Module take ownership and immediately
delete the object.