Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> seems like you want to use scoped_ptr here

I debated about that. There are other pointers in the test where Module does take ownership and deletes in the normal way, and so they can't use scoped_ptr. It seems like the the structures should be identical in everything but the one difference the test is checking: ownership-taken/not-taken based on the bounds. 

So I don't think the distinction should be made until after the test completes.

That's my view anyway, I'm happy to change it if you think scoped_ptr for some not not others (eg, item2) is better.

Patch-set: 1