Update patch set 1
Patch Set 1:

> I debated about that. There are other pointers in the test where Module does take ownership and deletes in the normal way, and so they can't use scoped_ptr. It seems like the the structures should be identical in everything but the one difference the test is checking: ownership-taken/not-taken based on the bounds. 
> 
> So I don't think the distinction should be made until after the test completes.
> 
> That's my view anyway, I'm happy to change it if you think scoped_ptr for some not not others (eg, item2) is better.

it sounds like the Module API handles ownership poorly, and callers have no way of doing it correctly.

the API docs state (i'll focus on stack frames, but the critique applies equally to all):

  // Add STACK_FRAME_ENTRY to the module.
  // This module owns all StackFrameEntry objects added with this
  // function: destroying the module destroys them as well.
  void AddStackFrameEntry(StackFrameEntry *stack_frame_entry);

with the implementation:

  void Module::AddStackFrameEntry(StackFrameEntry *stack_frame_entry) {
    if (!AddressIsInModule(stack_frame_entry->address)) {
      return;
    }

    stack_frame_entries_.push_back(stack_frame_entry);
  }

which means we force callers to reimplement AddressIsInModule/etc... logic just so they can do:

 if (MyCustomCheckerAddressIsInModule(m, extern)) {
   // Module won't own the object because it won't add it.
   del extern;
 } else {
   m.AddStackFrameEntry(extern);
 }

seems like we should have Module assume ownership of all objects added to it, or we adjust the API so the caller knows for sure what happened.

for the former, it'd mean creating a sep list like stack_frame_entries_owned_ or something just so we can walk it in the deconstructor to delete.

for the latter, it'd be:

  // Add STACK_FRAME_ENTRY to the module.
  // This module owns all StackFrameEntry objects successfully added with this
  // function: destroying the module destroys them as well.
  bool AddStackFrameEntry(StackFrameEntry *stack_frame_entry);
 ...
  bool Module::AddStackFrameEntry(StackFrameEntry *stack_frame_entry) {
    if (!AddressIsInModule(stack_frame_entry->address)) {
      return false;
    }

    stack_frame_entries_.push_back(stack_frame_entry);
    return true;
  }

then the caller can write:

 if (!m.AddStackFrameEntry(entry))
   delete entry;

Patch-set: 1
Reviewer: Gerrit User 1000768 <1000768@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>