Update patch set 1
Patch Set 1:

so the reason i didn't suggest this route is because the API talks about taking ownership, and deleting the object immediately is possibly surprising.  so one could create an object, add it to the module, and then use it later on (perhaps in diagnostics or checks or whatever).  but this API change means once it's been added, the caller has no access to it.

basically this API means it now has move semantics.  is there a way we can define that in the API so it's clear (not just as comments) ?  this is hitting the limit of my C++ knowledge.

Patch-set: 1
CC: Gerrit User 1000768 <1000768@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>