Update patch set 1
Patch Set 1:

(13 comments)

gLinux/Debian still allow you to install automake-1.15 in parallel, so that doesn't seem like it should be hard to set up.

i don't mind pulling up to the latest automake in general, but i think there might be a bit of a skew with distros & automake 1.16/1.16.1/1.16.2 due to their recent releases this year, and most people can't handle autotools it seems.  so sticking with 1.15.1 as long as it's easy for people to install seems like the least bad option.

Patch-set: 1