[RFC]linux_dumper: Do not merge if mapping contains the ELF header
If multiple shared libraries are mapped from the same file (APK on Android),
we cannot merge an executable mapping with a preceding read-only mapping just
based on the filename. For example, these two mappings correspond to different
libraries:

```
7fa38cc000-7fa38cd000 rw-p 00439000 [...] /xxx.apk
7fa38cd000-7fa38ce000 r-xp 00609000 [...] /xxx.apk
```

`EnumerateMappings` now checks if a mapping contains the ELF magic (indicating
that it corresponds to the first `PT_LOAD` segment of a new shared library) and
avoids merging in that case.

Change-Id: Iad1355e83e5a0dc17586dc8ab1a6a8907bf4232d