autoreconf
tools/linux/pid2md: create minidump from live process


tools/linux/core_handler: handler for coredumps


elf_core_dump: add a fallback to /proc/<pid>/mem

Bug: This change propose a way to use breakpad as a core handler on
Change-Id: I9298d81159029cefb81c915831db54884310ad05
Linux (see core(5)). This developement is motivated by the need to
capture the system logs (and other informations) into the minidump.
Because of the way minidumps are generated on Linux, it is not
possible to use external libraries during the generation of the dump
(like sd-journal). A true out-of-process mecanism is needed. One way
is to convert coredumps into minidumps. Unfortunatly this is not
always possible on embedded systems because it can take to much time
and consume to much disk. The proposed handler uses /proc/<pid>/mem
and does not need to read the full coredump.