core_handler: coredump handler to produce minidump
On Linux, it is possible to register a core handler via
/proc/sys/kernel/core_pattern. Doing so invokes the core handler when
a process crash. The core_handler uses /proc/<pid>/mem to access the
process memory. This way it is not necessary to process the full
coredump which takes time and consumes memory.

In order to profit from this core handler, for example, one can
integrate dump_syms into Yocto and generate an archive with the
breakpad symbols of all the binaries in the rootfs. Minidumps are
especially useful on embedded systems since they are lightweight and
provide contextual information.

Bug: This development is motivated by the need to capture the system
logs (and other information) into the minidump. Because of the way
minidumps are generated on Linux, it is not possible to use external
libraries during the generation of the dump (like sd-journal(3)).
A true out-of-process mechanism is needed.

Change-Id: I9298d81159029cefb81c915831db54884310ad05