Fix macos_dump_syms when __eh_frame is included in the dSYM file.
macho_reader assumes that every section in a segment exists, or none of
the sections exist in the file (for dSYM files).

https://reviews.llvm.org/D94460?id=315965 added __eh_frame section to
the __TEXT segments in dSYM files. All the other sections are removed,
but still have non-zero size in the header. macho_reader only looks at
`fileoff` and `filesize` fields to determine the size of the segment,
but it looks at `addr` and `size` to determine the size of the section,
therefore it determines that the sections would not fit in the segment
and refused to parse the file.

In this case the removed sections all have offset == 0. Ignore such
cases.

Change-Id: Ida6273d9bf9c95bd92636ea2e060a6d4ca70247d