Don't delete duplicate functions
This is one solution for https://crbug.com/google-breakpad/843.
I believe the duplicate funtions instroduced by ICF, which usually are
small percentage of all functions. It's okay to not release the memory.

Another way to solve it could be using shared_ptr instead of Function*,
and use weak_ptr in forward_ref_die_to_func. That requies more changes.

Bug: google-breakpad:843
Change-Id: I9ebad953ab43545e501af62ca60b0aadbabda56f