Write correct CVRecord for PE modules in minidump_writer
Crashes produced from Wine may contain both ELF and PE modules.
When writing the minidump file we check if the module has a PE
format and in case it is, write MDCVInfoPDB70 record and not
MDCVInfoELF one. If the file cannot be opened, its size is
smaller than 2, or it doesn't start with MZ symbols, then fall
back to the ELF functionality.

Bug: 196817261
Change-Id: I3af9da1bb128a5013ac7768d07d00ed74f7b2079