Update buildsystem generated files again
Add a bunch of missing quotes to configure.ac


Update ax_pthread-m4 to latest upstream version

This resolves further warnings when generating the configure file
with recent autoreconf versions.


Update autoconf and require autoconf 2.69

This removes/replaces a bunch of obsolete autoconf macros, so that
the configure file generates without warnings with recent autoconf
versions.


Update generated autotools files

These were regenerated with autoreconf (GNU Autoconf) 2.71


Fix building in subfolder

When building in a subfolder instead of the repository
root, the -I flag would relatively point to ./src which is
incorrect in that case. Instead use an absolute path here.
Same for the test driver path.

Change-Id: I63bc3dd72e959af857d11f907fa6d27b65527300