Adpat to SIGSTKSZ type in glibc 2.34
This is a rebase of CL 3226470 which is stalled
for quite some time and we are preparing ChromeOS
to move to glibc 2.34 and beyond, so we need it.

<https://sourceware.org/git/?p=glibc.git;a=commit;h=6c57d320484988e87e446e2e60ce42816bf51d53>
"sysconf: Add _SC_MINSIGSTKSZ/_SC_SIGSTKSZ [BZ #20305]" introduced
sysdeps/unix/sysv/linux/bits/sigstksz.h containing

> # define SIGSTKSZ sysconf (_SC_SIGSTKSZ)

so that SIGSTKSZ would have type long, causing

> src/client/linux/handler/exception_handler.cc: In function ‘void google_breakpad::{anonymous}::InstallAlternateStackLocked()’:
> src/client/linux/handler/exception_handler.cc:141:49: error: no matching function for call to ‘max(int, long int)’
>   141 |   static const unsigned kSigStackSize = std::max(16384, SIGSTKSZ);
>       |                                         ~~~~~~~~^~~~~~~~~~~~~~~~~

when building ExternalProject_breakpad as part of LibreOffice (see
<https://git.libreoffice.org/core/+/c90cb1b3c265c1f81e614e47652fe24137d7928c%5E!/>
"external/breakpad: Adpat to SIGSTKSZ type in glibc 2.34").  At least SuS4
appears to be silent about a required type of SIGSTKSZ defined in signal.h.

BUG=b:235960683
TEST=Local builds against glibc 2.34.

Change-Id: Iec6ce3d6ca295fcf8387aa299e2df61dc1477b08