allow dump_syms to operate on contents in memory
dump_syms assumes it is operating on a file and can access a commpliant
file system. This change allows dump_syms to operate on the contents of
a file that has already been read into memory. This is useful in a
server context where the file does not exist on the local file system.

Change-Id: Id63f115c2df287083d548dadd5ac487f97bde057