Support PE modules in core files when running core2md
Core files generated from `wine` contain both ELF and PE modules. Module
format can be guessed by checking the file contents. If the module
corresponds to PE-file conditions (has specific fields set up as
described in https://code.google.com/archive/p/corkami/wikis/PE.wiki) we'll
create a MDCVInfoPDB70 record in the minidump for it, but if the file
cannot be opened, is too short or is not a PE file, we'll fall back to
ELF procedure.

Test: build core2md and use it to convert a core file into dmp, validate
Change-Id: I225ffeea3f582deed40ecdfe7ab77f5754e90cbe
that the generated dmp file can be opened.