Refactor split dwarf handling.
When we have skeleton compilation unit (DW_TAG_skeleton_unit) in a
binary file refers to the complete unit in a split dwarf file
(.dwo/.dwp file), we should use the split dwarf file's path in warning
reporting. Right now, it uses the original file (binary file) path in
warning report, which is incorrect.

For example, if we have chrome.debug which is the binary with skeleton
debug info and chrome.dwp which is the complete debug info and the debug
info in chrome.dwp has some incorrect reference, it will warn on
chrome.debug rather than chrome.dwp.

This is also a preparation for a follow-up fix on relative offset being
inserted into the global inline_origin_map, which causes keys collision.

Bug: b/280290608
Change-Id: I9874bbd215b0054771f4829253c0f195ac52a5ea