Mac: truncate sym file names in upload_system_symbols
macOS caps filenames at 255 characters. When upload_system_symbols runs
`dump_syms`, the resulting filename is based on a mangled version of
the file's full path. In some circumstances (for example, the dumped
file itself lives in a temp directory), this name can exceed the max.

This change truncates the name from the end with some fudge factor.

Bug: 1400770
Change-Id: I68203a98eda2912893c5d8f7c676faee17e39e91