Mac: teach upload_system_symbols to extract installers and IPSWs
This change introduces two new flags: `--ipsw` and `--installer` which are mutually exclusive
with `--system-root` and each other. Each takes a file path as an argument, which is expected
to be an IPSW for `--ipsw` and an Apple installer for `--installer`.

Calling `upload_system_symbols` with these arguments will cause it to find any dyld shared
caches present inside the given IPSW or installer, extract them via `dsc_extractor` in
`--breakpad-tools`, then behave as if it had been called with the resulting libraries as
`--system-root`.
Bug: chromium:1400770
Change-Id: I7f98e0c6ab069a2e960f12773d800d8a5a37221f