linux_ptrace_dumper: Handle ARMEABI binaries on ARM64 kernels
For ARM32 binaries running on ARM64 Linux (kernel) reading of FP
registers fails. It's guarded for Android as well as softap platforms,
but other ARM platforms can suffer from this issue.

The ARMEABI linux_ptrace_dumper_unittest fails on system that runs
under ARM64 kernel.

In order to mitigate the issue, we adding a VFP registers read.
The Breakpad does not support include of VFP registers into a minidump
file, so that read is noop for the backend, but just a fix for broken
systems.

Bug: internal 322205293
Test: Run linux_ptrace_dumper_unittest on following [userland:kernel]
      combinations:
        armeabi-hardfp:armeabi
        armeabi-softfp:armeabi
        armeabi-hardfp:aarch64
        aarch64:aarch64

Signed-off-by: Volodymyr Riazantsev <riazantsevv@google.com>
Change-Id: I0709ae9a7ff913340ebc89de703ab2cb9c823b14